Add files via upload
## Catty Coin
The Catty Coin is here for revenge🐱

Dogs' time is up!

Catty is here for revenge and to reclaim his leadership over the internet.

Catty is the first token of a large collection to be launched this week, all part of the Catty Army.

Catty is an open source project entirely driven by the community. Whoever joins the Catty army enjoys the following benefits:

1-) Auto & Manual Burns; 

We have burned 50% of the supply. That means out of the 5% of the redistribution 50% will be distributed to the dead wallet. Coins are burned with every transaction. 🔥

2-) Locked Liquidty; 

Unruggable, majority of LP tokens locked & some burned so theres always LP. 🔒

3-) Rewards Holding;

 5% of each transaction is redistributed to token holders. That means you earn more $Catty Coins by just holding them in your wallet. 💎
 
4-) We have locked tokens 35% of the supply. 20% for 10 months, 15% for 24 months.🔒

So, %5 of each transaction added to Liquidity and %5 of each transaction redistributed to HODLers. 💎